fixed compilation warning
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@1595 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/media_tools/media_import.c
@@ -261,7 +261,7 @@ static GF_Err gf_import_still_image(GF_MediaImporter *import, Bool mult_desc_all
 	samp = gf_isom_sample_new();
 	samp->IsRAP = 1;
 	samp->dataLength = size;
-	if (import->initial_time_offset) samp->DTS = import->initial_time_offset*1000;
+	if (import->initial_time_offset) samp->DTS = (u64) (import->initial_time_offset*1000);
 
 	gf_import_message(import, GF_OK, "%s import %s - size %d x %d", (OTI==0x6C) ? "JPEG" : (OTI==0x6D) ? "PNG" : "JPEG2000", import->in_name, w, h);
